backend/pyproject.toml
[tool.poetry]
name = "ticketing-system"
version = "0.1.0"
description = ""
authors = ["Zaliant1 <ATMendoza1092@gmail.com>"]
packages = []

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.1.1"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pymongo = "^4.3.3"
python-dotenv = "^0.21.1"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
poethepoet = "^0.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks]
api = { shell = 'uvicorn src.ticketing_system.api.app:app --reload'}
bot = { script = 'src.ticketing_system.bot.main:main'}